Tidied code by spliting into components.
Made separate components for the top bar, the searchbar and the display grid. Got rid of all the testing pages now that components are done being tested. Started a "display" page which will include and introduction to the page and the content display grid and more things on the top navbar.